ajustado para criar o controler de auth
